GETs de ProductoJ se complementa con informacion de VendedorJ
Se implementa VendedorJ para agregar la informacion de Vendedor, Estudiante y Persona a los Productos

+++ API_MercaditoTEC/API_MercaditoTEC/Dtos/DtosJ/ProductoJ/ProductoJReadDto.cs
@@ -10,7 +10,7 @@ namespace API_MercaditoTEC.Dtos.DtosJ.ProductoJ
         [Required]
         public int idVendedor { get; set; }
         public string nombreVendedor { get; set; } //Persona //Estudiante //Vendedor
-        public int calificacionPromedioVendedor { get; set; } //Vendedor
+        public Single calificacionPromedioVendedor { get; set; } //Vendedor
         [Required]
         [MaxLength(150)]
         public string nombre { get; set; }
